librdb/cache/pwl/ssd: fix m_bytes_allocated exceeding m_bytes_allocated_cap
Fixes: https://tracker.ceph.com/issues/50670

Different from the mode of RWL counting free entries, SSD cache mode uses
the allocated bytes and the total capacity to control whether the
allocation is successful. The number of bytes allocated may be due to
the multi-thread exceeding the total capacity.

On the other hand, the number of bytes allocated should be less than
the total capacity, not equal to, in order to distinguish full cache
from empty cache..

Signed-off-by: Yin Congmin <congmin.yin@intel.com>